Extracting aam fields from order in Purchase events
Summary:
Adding two methods in MagentoDataHelper to extract automatic advanced matching fields from user session and orders.
The arrays returned by these classes are used in ServerSideHelper to set the userData field in Conversions API events, only when AAM is enabled and in the specified fields.
Because a new parameter was added to ServerSideHelper::send(), the before interceptor class was modified.
ServerSideHelperTest unit test is updated as well.

Reviewed By: HeyMultiverse

Differential Revision: D24811673

fbshipit-source-id: 40eae5c7