Set the `content_type` correctly for configurable products
In the case of Configurable Products, the `content_ids` of the ViewContent
event refers to the ID of the parent product. This ID is stored in the
`item_group_id` of the individual products. The `content_type` property
of the event should be set to `product_group` to reflect that.