Disable catalog incremental sync, when catalog is disabled
Summary:
Some clients who install our extensions are interested only in CAPI but catalog is also enabled by default.

There is an option to stop the catalog sync cron, but this does not stop the extension from pushing products to the catalog during upload.

With this diff, if the catalog sync is turned off, even incremental pushes will be disabled.

Reviewed By: mrharel

Differential Revision: D32062333

fbshipit-source-id: 3cec74af